[move] : move to dir
