+++ src/components/myPage/ClubList.jsx
@@ -1,12 +1,11 @@
 import React from 'react';
-import { STSection } from '../MyPage';
 import ClubInfo from './ClubInfo';
 import { LiaUserFriendsSolid } from 'react-icons/lia';
+import { STSection } from './MyPage';
 
 const ClubList = () => {
   return (
     <STSection>
-      {/* 이모티콘 */}
       <h3 className="flex gap-2 border-b-2 border-slate-300 mt-2 ml-4 mr-4">
         <LiaUserFriendsSolid />
         신청한 모임
