+++ src/router/Router.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
-import MyPage from '../pages/myPages/MyPage';
+import MyPage from '../components/myPage/MyPage';
 
 export default function Router() {
   return (
@@ -11,8 +11,8 @@ export default function Router() {
           <Route path="/login" element={<Login />} />
           <Route path="/join" element={<Join />} /> */}
         <Route path="/mypage" element={<MyPage />} />
-        {/* <Route path="/detail/:id" element={<DetailedPost />} /> */}
-        {/* </Route> */}
+        {/* <Route path="/detail/:id" element={<DetailedPost />} />
+        </Route> */}
       </Routes>
     </BrowserRouter>
   );
